src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/drop-event-for-input-in-shadow-expected.txt

CONSOLE ERROR: line 21: Uncaught TypeError: host.createShadowRoot is not a function
PASS successfullyParsed is true

TEST COMPLETE
